Refactoring: function to create RegionalTaskRefactoring: moving excess code from R5RCore to other classes - Abstracting elements of TravelTimeMatrixComputer and DetailedItineraryPlanner to super classes
